refactor(T10133): decompose tasks.ts renderers into core/render/tasks (B8)
Migrates renderShow / renderList / renderFind / renderAdd / renderUpdate /
renderComplete / renderDelete / renderArchive / renderRestore (371 LOC)
from packages/cleo/src/cli/renderers/tasks.ts to
packages/core/src/render/tasks/. Each renderer self-registers into the B5
registry on module load under `(command, 'generic')`.

Zero behavior change — the existing `(data, quiet)` signature is preserved
and the dispatcher continues to invoke renderers directly until the
typed-envelope migration completes. Adds `RenderOptions.quiet` so registry
adapters can forward the legacy flag through `renderEnvelopeForHuman`.

Closes T10133
Epic: T10114
ADR: adr-077-human-render-contract
